backend/requirements.txt
# ── Core Framework ───────────────────────────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0

# ── Database ─────────────────────────────────────────────────────────────────
pymongo==4.7.2

# ── Security ─────────────────────────────────────────────────────────────────
bcrypt==4.1.3
python-dotenv==1.0.1

# ── AI / NLP ─────────────────────────────────────────────────────────────────
openai==1.30.1
# spaCy: install separately with pre-built wheels (see README)
# Run: pip install spacy --prefer-binary
# Then: python -m spacy download en_core_web_sm

# ── Utilities ────────────────────────────────────────────────────────────────
# gunicorn is Linux-only; use waitress on Windows for production:
waitress==3.0.0
